+++ NEWS.md
@@ -1,5 +1,7 @@
 # meditate 0.1.2.9000
 
+- In `ReadSessions`, fix bug with current streak not reporting correctly.
+
 - Add `PlotMandala`, draws a colored mandala.
 
 # meditate 0.1.2